requirements.txt
pandas
scikit-learn
numpy
matplotlib
seaborn
jupyter
chromadb
sentence-transformers
python-dotenv
fastapi
langchain